Add preliminary steps to integrate exact solutions over curved boundaries.
Create framework to define the integrand with respect to x
for advection-diffusion exact solutions.
This integrand is used to integrate the solution over
domains with curved boundaries.


Former-commit-id: 1bd8dfb38734d39d19090ca421bbdae98ae7b0e0

+++ src_2D/HighOrderReconstruction/HighOrder2D_Input.h
@@ -28,7 +28,7 @@ public:
 
   //! @name Field access
   //@{
-  static const short int & MaximumReconstructionOrder(void) { return MaxReconstructionOrder; }
+  static const int & MaximumReconstructionOrder(void) { return MaxReconstructionOrder; }
   //@}
 
   //! @name Functions for input-output and broadcast
